+++ backend/src/handler/tila/nahtavillaoloTilaManager.ts
@@ -15,6 +15,9 @@ async function createNahtavillaoloVaihePDF(
   projekti: DBProjekti,
   kieli: Kieli
 ) {
+  if (!julkaisu.kuulutusPaiva) {
+    throw new Error("julkaisu.kuulutusPaiva puuttuu");
+  }
   const pdf = await asiakirjaService.createNahtavillaoloKuulutusPdf({
     asiakirjaTyyppi,
     projekti,
@@ -89,8 +92,10 @@ class NahtavillaoloTilaManager extends TilaManager {
 
     const nahtavillaoloVaihe = getNahtavillaoloVaihe(projekti);
     nahtavillaoloVaihe.palautusSyy = syy;
-
-    await this.deletePDFs(projekti.oid, nahtavillaoloVaihe.nahtavillaoloPDFt);
+    if (!julkaisuWaitingForApproval.nahtavillaoloPDFt) {
+      throw new Error("julkaisuWaitingForApproval.nahtavillaoloPDFt puuttuu");
+    }
+    await this.deletePDFs(projekti.oid, julkaisuWaitingForApproval.nahtavillaoloPDFt);
 
     await projektiDatabase.saveProjekti({ oid: projekti.oid, nahtavillaoloVaihe });
     await projektiDatabase.deleteNahtavillaoloVaiheJulkaisu(projekti, julkaisuWaitingForApproval.id);
@@ -119,7 +124,7 @@ class NahtavillaoloTilaManager extends TilaManager {
       return { nahtavillaoloPDFPath, nahtavillaoloIlmoitusPDFPath, nahtavillaoloIlmoitusKiinteistonOmistajallePDFPath };
     }
 
-    const pdfs = {};
+    const pdfs: LocalizedMap<NahtavillaoloPDF> = {};
     pdfs[kielitiedot.ensisijainenKieli] = await generatePDFsForLanguage(kielitiedot.ensisijainenKieli, julkaisuWaitingForApproval);
 
     if (kielitiedot.toissijainenKieli) {
@@ -130,6 +135,9 @@ class NahtavillaoloTilaManager extends TilaManager {
 
   private async deletePDFs(oid: string, nahtavillaoloPDFt: LocalizedMap<NahtavillaoloPDF>) {
     for (const language in nahtavillaoloPDFt) {
+      // nahtavillaoloPDFt ei ole null, ja language on tyyppiä Kieli, joka on nahtavillaoloPDFt:n avain
+      // eslint-disable-next-line @typescript-eslint/ban-ts-comment
+      // @ts-ignore
       const pdfs: NahtavillaoloPDF = nahtavillaoloPDFt[language];
       await fileService.deleteYllapitoFileFromProjekti({
         oid,
